+++ src/NCalc/Exceptions/NCalcEvaluationException.cs
@@ -1,14 +1,3 @@
 ﻿namespace NCalc.Exceptions;
 
-public sealed class NCalcEvaluationException : NCalcException
-{
-    public NCalcEvaluationException(string message)
-        : base(message)
-    {
-    }
-
-    public NCalcEvaluationException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
-}
+public class NCalcEvaluationException(string message) : NCalcException(message);