+++ src/NCalc/Factories/IExpressionFactory.cs
@@ -1,0 +1,14 @@
+﻿using NCalc.Domain;
+
+namespace NCalc.Factories;
+
+public interface IExpressionFactory
+{
+    public Expression Create(
+        string expression,
+        ExpressionContext? expressionContext = null);
+    
+    public Expression Create(
+        LogicalExpression logicalExpression,
+        ExpressionContext? expressionContext = null);
+}