+++ src/NCalc/Handlers/EvaluateFunctionHandler.cs
@@ -1,5 +1,3 @@
-﻿using NCalc.Domain;
-
-namespace NCalc.Handlers;
+﻿namespace NCalc.Handlers;
 
 public delegate void EvaluateFunctionHandler(string name, FunctionArgs args, string? id = null);