+++ src/NCalc/Handlers/EvaluateParameterHandler.cs
@@ -1,5 +1,3 @@
-﻿using NCalc.Domain;
-
-namespace NCalc.Handlers;
+﻿namespace NCalc.Handlers;
 
 public delegate void EvaluateParameterHandler(string name, ParameterArgs args);