+++ src/NCalc/Visitors/EvaluationVisitor.cs
@@ -1,39 +1,41 @@
 ﻿using NCalc.Domain;
+using NCalc.Exceptions;
 using NCalc.Handlers;
 using NCalc.Helpers;
 
 namespace NCalc.Visitors;
 
-public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInfo) : LogicalExpressionVisitor
+public class EvaluationVisitor : IEvaluationVisitor
 {
     public event EvaluateFunctionHandler? EvaluateFunction;
     public event EvaluateParameterHandler? EvaluateParameter;
     
-    public ExpressionOptions Options { get; set; } = options;
-    public CultureInfo CultureInfo { get; set; } = cultureInfo;
-
-    public required Dictionary<string, object?> Parameters { get; set; }
-
-    public object? Result { get; private set; }
+    public ExpressionOptions Options
+    {
+        get => Context.Options;
+        set => Context.Options = value;
+    }
     
-    private static readonly Type[] BuiltInTypes =
-    [
-        typeof(decimal),
-        typeof(double),
-        typeof(float),
-        typeof(long),
-        typeof(ulong),
-        typeof(int),
-        typeof(uint),
-        typeof(short),
-        typeof(ushort),
-        typeof(byte),
-        typeof(sbyte),
-        typeof(char),
-        typeof(bool),
-        typeof(string),
-        typeof(object)
-    ];
+    public CultureInfo CultureInfo
+    {
+        get => Context.CultureInfo;
+        set => Context.CultureInfo = value;
+    }
+    
+    public Dictionary<string, object?> Parameters { get; set; } = new();
+
+    public ExpressionContext Context { get; set; } = new();
+
+    public object? Result { get; set; }
+
+    public EvaluationVisitor()
+    {
+    }
+
+    public EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInfo)
+    {
+        Context = new(options, cultureInfo);
+    }
     
     private object? Evaluate(LogicalExpression expression)
     {
@@ -41,50 +43,12 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
         return Result;
     }
 
-    public override void Visit(LogicalExpression expression)
+    public void Visit(LogicalExpression expression)
     {
         throw new NotSupportedException("The Visit method is not supported for this class.");
     }
 
-    /// <summary>
-    /// Gets the the most precise type.
-    /// </summary>
-    /// <param name="a">Type a.</param>
-    /// <param name="b">Type b.</param>
-    /// <returns></returns>
-    private static Type GetMostPreciseType(Type? a, Type? b)
-    {
-        foreach (var t in BuiltInTypes)
-        {
-            if (a == t || b == t)
-            {
-                return t;
-            }
-        }
-
-        return a ?? typeof(object);
-    }
-
-    public int CompareUsingMostPreciseType(object? a, object? b)
-    {
-        Type mpt;
-        if (a == null)
-            mpt = GetMostPreciseType(null, b?.GetType());
-        else
-            mpt = GetMostPreciseType(a.GetType(), b?.GetType());
-        
-        var aValue = a != null ? Convert.ChangeType(a, mpt, CultureInfo) : null;
-        var bValue = b != null ? Convert.ChangeType(b, mpt, CultureInfo) : null;
-
-        bool isCaseInsensitiveComparer = Options.HasOption(ExpressionOptions.CaseInsensitiveComparer);
-
-        if (isCaseInsensitiveComparer)
-            return CaseInsensitiveComparer.Default.Compare(aValue, bValue);
-
-        return Comparer.Default.Compare(aValue, bValue);
-    }
-
-    public override void Visit(TernaryExpression expression)
+    public void Visit(TernaryExpression expression)
     {
         // Evaluates the left expression and saves the value
         expression.LeftExpression.Accept(this);
@@ -99,10 +63,8 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             expression.RightExpression.Accept(this);
         }
     }
-
-    private static bool IsReal(object? value) => value is decimal or double or float;
-
-    public override void Visit(BinaryExpression expression)
+    
+    public void Visit(BinaryExpression expression)
     {
         // simulate Lazy<Func<>> behavior for late evaluation
         object? leftValue = null;
@@ -121,7 +83,7 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
                 break;
 
             case BinaryExpressionType.Div:
-                Result = IsReal(Left()) || IsReal(Right())
+                Result = TypeHelper.IsReal(Left()) || TypeHelper.IsReal(Right())
                     ? MathHelper.Divide(Left(), Right(), CultureInfo)
                     : MathHelper.Divide(Convert.ToDouble(Left(), CultureInfo), Right(), CultureInfo);
                 break;
@@ -230,7 +192,7 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
         }
     }
 
-    public override void Visit(UnaryExpression expression)
+    public void Visit(UnaryExpression expression)
     {
         // Recursively evaluates the underlying expression
         expression.Expression.Accept(this);
@@ -255,12 +217,13 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
         }
     }
 
-    public override void Visit(ValueExpression expression)
+    public void Visit(ValueExpression expression)
     {
         Result = expression.Value;
     }
-    
-    public override void Visit(Function function)
+
+
+    public void Visit(Function function)
     {
         var args = new FunctionArgs
         {
@@ -280,10 +243,9 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             args.Parameters[i].Parameters = Parameters;
         }
 
-        bool ignoreCase = Options.HasOption(ExpressionOptions.IgnoreCase);
-
         // Calls external implementation
-        OnEvaluateFunction(ignoreCase ? function.Identifier.Name.ToLower() : function.Identifier.Name, args, function.Id.ToString());
+
+        OnEvaluateFunction(function.Identifier.Name, args, function.Id.ToString());
 
         // If an external implementation was found get the result back
         if (args.HasResult)
@@ -304,35 +266,35 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "ABS":
                 CheckCase("Abs", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Abs() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Abs() takes exactly 1 argument");
                 Result = Math.Abs(Convert.ToDecimal(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "ACOS":
                 CheckCase("Acos", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Acos() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Acos() takes exactly 1 argument");
                 Result = Math.Acos(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "ASIN":
                 CheckCase("Asin", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Asin() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Asin() takes exactly 1 argument");
                 Result = Math.Asin(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "ATAN":
                 CheckCase("Atan", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Atan() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Atan() takes exactly 1 argument");
                 Result = Math.Atan(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "ATAN2":
                 CheckCase("Atan2", function.Identifier.Name);
                 if (function.Expressions.Length != 2)
-                    throw new ArgumentException("Atan2() takes exactly 2 argument");
+                    throw new NCalcEvaluationException("Atan2() takes exactly 2 argument");
                 Result = Math.Atan2(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo),
                     Convert.ToDouble(Evaluate(function.Expressions[1]), CultureInfo));
                 break;
@@ -340,35 +302,35 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "CEILING":
                 CheckCase("Ceiling", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Ceiling() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Ceiling() takes exactly 1 argument");
                 Result = Math.Ceiling(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "COS":
                 CheckCase("Cos", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Cos() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Cos() takes exactly 1 argument");
                 Result = Math.Cos(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "EXP":
                 CheckCase("Exp", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Exp() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Exp() takes exactly 1 argument");
                 Result = Math.Exp(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "FLOOR":
                 CheckCase("Floor", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Floor() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Floor() takes exactly 1 argument");
                 Result = Math.Floor(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "IEEEREMAINDER":
                 CheckCase("IEEERemainder", function.Identifier.Name);
                 if (function.Expressions.Length != 2)
-                    throw new ArgumentException("IEEERemainder() takes exactly 2 arguments");
+                    throw new NCalcEvaluationException("IEEERemainder() takes exactly 2 arguments");
                 Result = Math.IEEERemainder(Convert.ToDouble(Evaluate(function.Expressions[0])),
                     Convert.ToDouble(Evaluate(function.Expressions[1]), CultureInfo));
                 break;
@@ -376,14 +338,14 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "LN":
                 CheckCase("Ln", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Ln() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Ln() takes exactly 1 argument");
                 Result = Math.Log(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "LOG":
                 CheckCase("Log", function.Identifier.Name);
                 if (function.Expressions.Length != 2)
-                    throw new ArgumentException("Log() takes exactly 2 arguments");
+                    throw new NCalcEvaluationException("Log() takes exactly 2 arguments");
                 Result = Math.Log(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo),
                     Convert.ToDouble(Evaluate(function.Expressions[1]), CultureInfo));
                 break;
@@ -391,14 +353,14 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "LOG10":
                 CheckCase("Log10", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Log10() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Log10() takes exactly 1 argument");
                 Result = Math.Log10(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "POW":
                 CheckCase("Pow", function.Identifier.Name);
                 if (function.Expressions.Length != 2)
-                    throw new ArgumentException("Pow() takes exactly 2 arguments");
+                    throw new NCalcEvaluationException("Pow() takes exactly 2 arguments");
                 Result = Math.Pow(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo),
                     Convert.ToDouble(Evaluate(function.Expressions[1]), CultureInfo));
                 break;
@@ -406,7 +368,7 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "ROUND":
                 CheckCase("Round", function.Identifier.Name);
                 if (function.Expressions.Length != 2)
-                    throw new ArgumentException("Round() takes exactly 2 arguments");
+                    throw new NCalcEvaluationException("Round() takes exactly 2 arguments");
 
                 var rounding = (Options & ExpressionOptions.RoundAwayFromZero) == ExpressionOptions.RoundAwayFromZero ? MidpointRounding.AwayFromZero : MidpointRounding.ToEven;
 
@@ -417,42 +379,42 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "SIGN":
                 CheckCase("Sign", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Sign() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Sign() takes exactly 1 argument");
                 Result = Math.Sign(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "SIN":
                 CheckCase("Sin", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Sin() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Sin() takes exactly 1 argument");
                 Result = Math.Sin(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "SQRT":
                 CheckCase("Sqrt", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Sqrt() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Sqrt() takes exactly 1 argument");
                 Result = Math.Sqrt(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "TAN":
                 CheckCase("Tan", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Tan() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Tan() takes exactly 1 argument");
                 Result = Math.Tan(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "TRUNCATE":
                 CheckCase("Truncate", function.Identifier.Name);
                 if (function.Expressions.Length != 1)
-                    throw new ArgumentException("Truncate() takes exactly 1 argument");
+                    throw new NCalcEvaluationException("Truncate() takes exactly 1 argument");
                 Result = Math.Truncate(Convert.ToDouble(Evaluate(function.Expressions[0]), CultureInfo));
                 break;
 
             case "MAX":
                 CheckCase("Max", function.Identifier.Name);
                 if (function.Expressions.Length != 2)
-                    throw new ArgumentException("Max() takes exactly 2 arguments");
+                    throw new NCalcEvaluationException("Max() takes exactly 2 arguments");
                 var maxleft = Evaluate(function.Expressions[0]);
                 var maxright = Evaluate(function.Expressions[1]);
                 Result = MathHelper.Max(maxleft, maxright, CultureInfo);
@@ -461,7 +423,7 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "MIN":
                 CheckCase("Min", function.Identifier.Name);
                 if (function.Expressions.Length != 2)
-                    throw new ArgumentException("Min() takes exactly 2 arguments");
+                    throw new NCalcEvaluationException("Min() takes exactly 2 arguments");
                 var minleft = Evaluate(function.Expressions[0]);
                 var minright = Evaluate(function.Expressions[1]);
                 Result = MathHelper.Min(minleft, minright, CultureInfo);
@@ -470,7 +432,7 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "IFS":
                 CheckCase("ifs", function.Identifier.Name);
                 if (function.Expressions.Length < 3 || function.Expressions.Length % 2 != 1)
-                    throw new ArgumentException("ifs() takes at least 3 arguments, or an odd number of arguments");
+                    throw new NCalcEvaluationException("ifs() takes at least 3 arguments, or an odd number of arguments");
                 foreach (var eval in function.Expressions.Where((_, i) => i % 2 == 0))
                 {
                     var index = Array.IndexOf(function.Expressions, eval);
@@ -489,7 +451,7 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "IF":
                 CheckCase("if", function.Identifier.Name);
                 if (function.Expressions.Length != 3)
-                    throw new ArgumentException("if() takes exactly 3 arguments");
+                    throw new NCalcEvaluationException("if() takes exactly 3 arguments");
                 var cond = Convert.ToBoolean(Evaluate(function.Expressions[0]), CultureInfo);
                 Result = cond ? Evaluate(function.Expressions[1]) : Evaluate(function.Expressions[2]);
                 break;
@@ -497,7 +459,7 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             case "IN":
                 CheckCase("in", function.Identifier.Name);
                 if (function.Expressions.Length < 2)
-                    throw new ArgumentException("in() takes at least 2 arguments");
+                    throw new NCalcEvaluationException("in() takes at least 2 arguments");
                 var parameter = Evaluate(function.Expressions[0]);
                 var evaluation = false;
                 for (var i = 1; i < function.Expressions.Length; i++)
@@ -514,33 +476,26 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
                 break;
 
             default:
-                throw new ArgumentException("Function not found", function.Identifier.Name);
+                throw new NCalcFunctionNotFoundException($"Function {function.Identifier.Name} not found", function.Identifier.Name);
         }
+    }
+
+    public int CompareUsingMostPreciseType(object? a, object? b)
+    {
+        return TypeHelper.CompareUsingMostPreciseType(a, b,new(CultureInfo, Options.HasOption(ExpressionOptions.CaseInsensitiveComparer)));
     }
 
     private void CheckCase(string function, string called)
     {
         bool ignoreCase = Options.HasOption(ExpressionOptions.IgnoreCase);
 
-        if (ignoreCase)
-        {
-            if (function.Equals(called, StringComparison.OrdinalIgnoreCase))
-            {
-                return;
-            }
-
-            throw new ArgumentException("Function not found", called);
-        }
-
-        if (function != called)
-        {
-            throw new ArgumentException($"Function not found {called}. Try {function} instead.");
-        }
+        if (!ignoreCase && function != called)
+            throw new NCalcFunctionNotFoundException($"Function {called} not found. Try {function} instead.", called);
     }
     
-    public override void Visit(Identifier parameter)
+    public void Visit(Identifier identifier)
     {
-        if (Parameters.TryGetValue(parameter.Name, out var parameterValue))
+        if (Parameters.TryGetValue(identifier.Name, out var parameterValue))
         {
             // The parameter is defined in the hashtable
             if (parameterValue is Expression expression)
@@ -548,7 +503,8 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
                 // Overloads parameters 
                 foreach (var p in Parameters)
                 {
-                    expression.Parameters[p.Key] = p.Value;
+                    if (expression.Parameters != null) 
+                        expression.Parameters[p.Key] = p.Value;
                 }
 
                 expression.EvaluateFunction += EvaluateFunction;
@@ -565,10 +521,10 @@ public class EvaluationVisitor(ExpressionOptions options, CultureInfo cultureInf
             var args = new ParameterArgs();
 
             // Calls external implementation
-            OnEvaluateParameter(parameter.Name, args);
+            OnEvaluateParameter(identifier.Name, args);
 
             if (!args.HasResult)
-                throw new ArgumentException("Parameter was not defined", parameter.Name);
+                throw new NCalcParameterNotDefinedException(identifier.Name);
 
             Result = args.Result;
         }